AMP 1.14.11 merged into AMP 1.15
git-svn-id: file:///svn/branches/AMP_1_15_RELEASE/amp@13569 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/dgfoundation/amp/ar/view/xls/ReportHeadingsXLS.java
@@ -173,8 +173,8 @@ public class ReportHeadingsXLS extends XLSExporter {
 					if (!"-".equalsIgnoreCase(col.getName(metadata.getHideActivities()))){
 						if(col.getWidth()>1) {
 							for(int k=0;k<col.getWidth();k++) {
-								HSSFCell cell = row.getCell(colId.intValue()+k);
-								if(cell==null) cell=row.createCell(colId.intValue()+k);
+								HSSFCell cell = row.getCell((short) (colId.intValue()+k));
+								if(cell==null) cell=row.createCell((short) (colId.intValue()+k));
 								HSSFCellStyle cellstyle = wb.createCellStyle();
 								cellstyle.cloneStyleFrom(this.getHighlightedStyle());
 								cell.setCellStyle(cellstyle);
