+++ WEB-INF/src/org/dgfoundation/amp/ar/view/xls/XLSExporter.java
@@ -98,7 +98,7 @@ public abstract class XLSExporter extends Exporter {
 	}
 
 	protected HSSFCell getCell(HSSFRow row, HSSFCellStyle style) {
-		HSSFCell cell = row.createCell(colId.intValue());
+		HSSFCell cell = row.createCell((short)colId.intValue());
 //		HSSFCellStyle cellstyle = wb.createCellStyle();
 //		cellstyle.cloneStyleFrom(style);
 		cell.setCellStyle(style);
